requirements.txt
python-dateutil
feedparser
beautifulsoup4
py-postgresql
pymongo
